Add files via upload
These are the files uploaded for the first two assignments of the AI In Industry Course.
The first assignment describes the client-server architecture and the second one describes the  YOLO model trained with the Coco dataset to detect people. The parameters are then set and the object detection takes place to label them into boxes whilst displaying the detected output.